Do not show drag buttons on elements that are only children
Since dragging elements is for sorting elements among its siblings, this button should not be seen on only children.

One possible are of confusion for this is when an element, like an `<strong>`, is placed in the middle of Text nodes that aren't wrapped in any tag. In that situation dragging isn't allowed either because unlike with DOM elements, there's no good way of telling where we want that element to land (in between any two characters?)