Fix a race condition between updating the AST from client/server and saving the changes made on the template.
Since we have 2 distinct events: one to update the AST and another
to actually save the changes, on an env with high latency
when one makes a change on the visual editor and hit Save Change
there could a race condition that would save the old template
because it wasn't updated yet.

So we need to use another assign that is guaranteed to be updated
before the save event actually happens.