Fixes the calculations of the positions when pages have no wrapper
When the elements of the layout and the elements of the page share the same wrapper, positioning calculations are broken.

Adds a wrapper with `display: contents` that doesn't have any effect on CSS, but still allows to tell apart which items belong to the page and which ones belong to the surounding layout.