Allow to drop elements into the page over the layout
While the elements cannot be dropped inside the layout (we'd need to be editing the layout itself to allow changing it), they are not valid drop targets.

If an element is dropped over an element of the layout before the main content, it become the first element on the page
Conversely, it will dropped at the very end of the page if it's dropped after  the page content.